third_party/WebKit/LayoutTests/fast/css/atapply/multiple-at-apply-rules.html

<!doctype html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<style>
#e {
    --foo: {
        background-color: red;
        width: 100px;
        height: 200px;
    };
    --bar: {
        background-color: green;
    };
    @apply --foo;
    @apply --bar;
}
</style>
<div id=e>
</div>
<script>
test(function(){
    assert_equals(getComputedStyle(e).backgroundColor, "rgb(0, 128, 0)");
    assert_equals(getComputedStyle(e).width, "100px");
    assert_equals(getComputedStyle(e).height, "200px");
}, "Multiple @apply rules can be used simultaneously");
</script>
